backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pgvector==0.3.2
sentence-transformers==3.1.1
groq>=0.13.0
redis==5.1.0
prometheus-client==0.21.0
pydantic-settings==2.5.2
python-multipart==0.0.12
